onMessage listeners fulfill the contract and return a promise
Previous versions used callbacks, which are still supported, but Promises are the way forward => each onMessage listener needs to return a promise (xor call the callback), message senders can await on the response.

This fixes "The message port closed before a response was received" we got most likely from the `ping` message, which used a callback but the corresponding listener did not call it.